requirements.txt
deepgram_sdk==3.2.7
openai==1.24.0
python-dotenv==1.0.1
supabase==2.4.3
